+++ bill/tax.go
@@ -23,15 +23,36 @@ type Tax struct {
 	// the rounding model used.
 	Rounding cbc.Key `json:"rounding,omitempty" jsonschema:"title=Rounding Model"`
 
+	// Point is a code that identifies the event which triggers the tax liability,
+	// such as invoice issuance, delivery of goods, or receipt of payment.
+	Point cbc.Key `json:"point,omitempty" jsonschema:"title=Point"`
+
 	// Additional extensions that are applied to the invoice as a whole as opposed to specific
 	// sections.
 	Ext tax.Extensions `json:"ext,omitempty" jsonschema:"title=Extensions"`
+
+	// Notes contains tax-related notes, typically used for exemption reasons
+	// or other tax-specific explanations associated with particular tax categories.
+	Notes []*tax.Note `json:"notes,omitempty" jsonschema:"title=Notes"`
 
 	// Any additional data that may be required for processing, but should never
 	// be relied upon by recipients.
 	Meta cbc.Meta `json:"meta,omitempty" jsonschema:"title=Meta"`
 
 	tags []cbc.Key
+}
+
+// MergeNotes adds a tax note to the tax object, automatically handling nil data,
+// and returning a new updated instance.
+func (t *Tax) MergeNotes(notes ...*tax.Note) *Tax {
+	if len(notes) == 0 {
+		return t
+	}
+	if t == nil {
+		t = new(Tax)
+	}
+	t.Notes = append(t.Notes, notes...)
+	return t
 }
 
 // MergeExtensions makes it easier to add extensions to the tax object
@@ -79,6 +100,7 @@ func (t *Tax) Normalize(normalizers tax.Normalizers) {
 		t.Rounding = tax.RoundingRuleCurrency
 	}
 	t.Ext = tax.CleanExtensions(t.Ext)
+	tax.Normalize(normalizers, t.Notes)
 	normalizers.Each(t)
 }
 
@@ -87,6 +109,11 @@ func taxRules() *rules.Set {
 		rules.Field("rounding",
 			rules.AssertIfPresent("01", "rounding model is not valid",
 				cbc.InKeyDefs(tax.RoundingRules),
+			),
+		),
+		rules.Field("point",
+			rules.AssertIfPresent("02", "tax point is not valid",
+				cbc.InKeyDefs(tax.PointDefs),
 			),
 		),
 	)
@@ -117,6 +144,16 @@ func (t Tax) JSONSchemaExtend(schema *jsonschema.Schema) {
 				Const:       r.Key.String(),
 				Title:       r.Name.String(),
 				Description: r.Desc.String(),
+			}
+		}
+	}
+	if p, ok := schema.Properties.Get("point"); ok {
+		p.OneOf = make([]*jsonschema.Schema, len(tax.PointDefs))
+		for i, w := range tax.PointDefs {
+			p.OneOf[i] = &jsonschema.Schema{
+				Const:       w.Key.String(),
+				Title:       w.Name.String(),
+				Description: w.Desc.String(),
 			}
 		}
 	}